+++ projects/igniteui-angular/src/lib/grids/common/pipes.ts
@@ -265,7 +265,7 @@ export class IgxStringReplacePipe implements PipeTransform {
 @Pipe({ name: 'transactionState' })
 export class IgxGridTransactionStatePipe implements PipeTransform {
 
-    transform(row_id: any, field: string, rowEditable: boolean, transactions: any, _: any) {
+    transform(row_id: any, field: string, rowEditable: boolean, transactions: any, _: any, __: any) {
         if (rowEditable) {
             const rowCurrentState = transactions.getAggregatedValue(row_id, false);
             if (rowCurrentState) {
@@ -277,5 +277,13 @@ export class IgxGridTransactionStatePipe implements PipeTransform {
             const value = resolveNestedPath(transaction?.value ?? {}, field);
             return transaction && transaction.value && (value || value === 0 || value === false);
         }
+    }
+}
+
+@Pipe({ name: 'columnFormatter' })
+export class IgxColumnFormatterPipe implements PipeTransform {
+
+    transform(value: any, formatter: (v: any) => any) {
+        return formatter(value);
     }
 }